+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SqlCommand.Xml.cs
@@ -13,6 +13,7 @@ using Microsoft.Data.ProviderBase;
 using Microsoft.Data.SqlClient.Connection;
 using Microsoft.Data.SqlClient.Server;
 using Microsoft.Data.SqlClient.Internal;
+using Microsoft.Data.SqlClient.Utilities;
 
 #if NETFRAMEWORK
 using System.Security.Permissions;
@@ -43,7 +44,7 @@ namespace Microsoft.Data.SqlClient
             #if NETFRAMEWORK
             SqlConnection.ExecutePermission.Demand();
             #endif
-            
+
             SqlClientEventSource.Log.TryCorrelationTraceEvent(
                 "SqlCommand.BeginExecuteXmlReader | API | Correlation | " +
                 $"Object Id {ObjectID}, " +
@@ -77,18 +78,18 @@ namespace Microsoft.Data.SqlClient
                     $"Command Text '{CommandText}'");
             }
         }
-        
+
         /// <include file='../../../../../../doc/snippets/Microsoft.Data.SqlClient/SqlCommand.xml' path='docs/members[@name="SqlCommand"]/ExecuteXmlReader/*'/>
         public XmlReader ExecuteXmlReader()
         {
             #if NETFRAMEWORK
             SqlConnection.ExecutePermission.Demand();
             #endif
-            
+
             // Reset _pendingCancel upon entry into any Execute - used to synchronize state
             // between entry into Execute* API and the thread obtaining the stateObject.
             _pendingCancel = false;
-            
+
             using var diagnosticScope = s_diagnosticListener.CreateCommandScope(this, _transaction);
 
             using var eventScope = SqlClientEventScope.Create($"SqlCommand.ExecuteXmlReader | API | Object Id {ObjectID}");
@@ -142,11 +143,11 @@ namespace Microsoft.Data.SqlClient
             IsProviderRetriable
                 ? InternalExecuteXmlReaderWithRetryAsync(cancellationToken)
                 : InternalExecuteXmlReaderAsync(cancellationToken);
-        
+
         #endregion
-        
+
         #region Private Methods
-        
+
         private static XmlReader CompleteXmlReader(SqlDataReader dataReader, bool isAsync)
         {
             XmlReader xmlReader = null;
@@ -189,7 +190,7 @@ namespace Microsoft.Data.SqlClient
                 $"Activity Id {ActivityCorrelator.Current}, " +
                 $"Client Connection Id {_activeConnection?.ClientConnectionId}, " +
                 $"Command Text '{CommandText}'");
-            
+
             return BeginExecuteXmlReaderInternal(
                 CommandBehavior.SequentialAccess,
                 callback,
@@ -215,7 +216,7 @@ namespace Microsoft.Data.SqlClient
                 // Reset _pendingCancel upon entry into any Execute - used to synchronize state
                 // between entry into Execute* API and the thread obtaining the stateObject.
                 _pendingCancel = false;
-                
+
                 // Special case - done outside of try/catches to prevent putting a stateObj back
                 // into pool when we should not.
                 ValidateAsyncCommand();
@@ -247,7 +248,7 @@ namespace Microsoft.Data.SqlClient
                         asyncWrite,
                         isRetry);
 
-                    // @TODO: NonQuery pathway has the continueTaskWithState block inside this try. One or the other seems wrong 
+                    // @TODO: NonQuery pathway has the continueTaskWithState block inside this try. One or the other seems wrong
                 }
                 catch (Exception e) when (ADP.IsCatchableOrSecurityExceptionType(e))
                 {
@@ -352,21 +353,21 @@ namespace Microsoft.Data.SqlClient
             if (task.IsFaulted)
             {
                 Exception e = task.Exception?.InnerException;
-                
+
                 s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
-                
+
                 source.SetException(e);
             }
             else if (task.IsCanceled)
             {
                 s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
-                
+
                 source.SetCanceled();
             }
             else
             {
                 s_diagnosticListener.WriteCommandAfter(operationId, this, _transaction);
-                
+
                 source.SetResult(task.Result);
             }
         }
@@ -374,7 +375,7 @@ namespace Microsoft.Data.SqlClient
         private XmlReader EndExecuteXmlReaderAsync(IAsyncResult asyncResult)
         {
             Debug.Assert(!_internalEndExecuteInitiated || _stateObj is null);
-            
+
             SqlClientEventSource.Log.TryCorrelationTraceEvent(
                 "SqlCommand.EndExecuteXmlReaderAsync | API | Correlation | " +
                 $"Object Id {ObjectID}, " +
@@ -389,9 +390,9 @@ namespace Microsoft.Data.SqlClient
                 ReliablePutStateObject();
                 throw asyncException.InnerException;
             }
-            
+
             ThrowIfReconnectionHasBeenCanceled();
-            
+
             // Locking _stateObj prevents races with close/cancel.
             // If we have already initiated the End call internally, we have already done that, so
             // no point doing it again.
@@ -405,7 +406,7 @@ namespace Microsoft.Data.SqlClient
 
             return EndExecuteXmlReaderInternal(asyncResult);
         }
-        
+
         private XmlReader EndExecuteXmlReaderInternal(IAsyncResult asyncResult)
         {
             bool success = false;
@@ -443,27 +444,27 @@ namespace Microsoft.Data.SqlClient
                 WriteEndExecuteEvent(success, sqlExceptionNumber, isSynchronous: false);
             }
         }
-        
+
         private Task<XmlReader> InternalExecuteXmlReaderAsync(CancellationToken cancellationToken)
         {
             #if NETFRAMEWORK
             SqlConnection.ExecutePermission.Demand();
             #endif
-            
+
             SqlClientEventSource.Log.TryCorrelationTraceEvent(
                 "SqlCommand.InternalExecuteXmlReaderAsync | API | Correlation | " +
                 $"Object Id {ObjectID}, " +
                 $"Activity Id {ActivityCorrelator.Current}, " +
                 $"Client Connection Id {_activeConnection?.ClientConnectionId}, " +
                 $"Command Text '{CommandText}'");
-            
+
             Guid operationId = s_diagnosticListener.WriteCommandBefore(this, _transaction);
-            
+
             // Connection can be used as state in RegisterForConnectionCloseNotification continuation
             // to avoid an allocation so use it as the state value if possible but it can be changed if
             // you need it for a more important piece of data that justifies the tuple allocation later
             TaskCompletionSource<XmlReader> source = new TaskCompletionSource<XmlReader>(_activeConnection);
-            
+
             CancellationTokenRegistration registration = new CancellationTokenRegistration();
             if (cancellationToken.CanBeCanceled)
             {
@@ -476,7 +477,7 @@ namespace Microsoft.Data.SqlClient
                 registration = cancellationToken.Register(callback: s_cancelIgnoreFailure, state: this);
             }
 
-            // @TODO: This can be cleaned up to lines if InnerConnection is always SqlInternalConnection 
+            // @TODO: This can be cleaned up to lines if InnerConnection is always SqlInternalConnection
             ExecuteXmlReaderAsyncCallContext context = null;
             if (_activeConnection?.InnerConnection is SqlConnectionInternal sqlInternalConnection)
             {
@@ -516,34 +517,34 @@ namespace Microsoft.Data.SqlClient
                     TaskCompletionSource<XmlReader> source = context.TaskCompletionSource;
 
                     context.Dispose();
-                        
+
                     command.CleanupAfterExecuteXmlReaderAsync(task, source, operationId);
                 },
                 scheduler: TaskScheduler.Default);
             }
-            catch (Exception e) 
+            catch (Exception e)
             {
                 s_diagnosticListener.WriteCommandError(operationId, this, _transaction, e);
-                
+
                 source.SetException(e);
             }
 
             return returnedTask;
         }
-        
+
         private Task<XmlReader> InternalExecuteXmlReaderWithRetryAsync(CancellationToken cancellationToken) =>
             RetryLogicProvider.ExecuteAsync(
                 sender: this,
                 () => InternalExecuteXmlReaderAsync(cancellationToken),
                 cancellationToken);
-        
+
         #endregion
 
         internal sealed class ExecuteXmlReaderAsyncCallContext
             : AAsyncCallContext<SqlCommand, XmlReader, CancellationTokenRegistration>
         {
             public SqlCommand Command => _owner;
-            
+
             public Guid OperationId { get; set; }
 
             public TaskCompletionSource<XmlReader> TaskCompletionSource => _source;
